Implemented build custom runtime

+++ lithops/serverless/backends/gcp_functions/gcp_functions.py
@@ -22,8 +22,6 @@ import httplib2
 import sys
 import zipfile
 import time
-import random
-import tempfile
 import textwrap
 import lithops
 from google.cloud import pubsub_v1
@@ -34,10 +32,11 @@ from googleapiclient.errors import HttpError
 from google.auth import jwt
 import google.api_core.exceptions
 
-from lithops.version import __version__
-from lithops.utils import version_str
-from lithops.storage import InternalStorage
 from . import config as gcp_config
+from ....version import __version__
+from ....utils import version_str
+from ....storage import InternalStorage
+from .... import config
 
 logger = logging.getLogger(__name__)
 logging.getLogger('googleapiclient').setLevel(logging.CRITICAL)
@@ -45,7 +44,7 @@ logging.getLogger('google_auth_httplib2').setLevel(logging.CRITICAL)
 logging.getLogger('google.auth.transport.requests').setLevel(logging.CRITICAL)
 logging.getLogger('google.cloud.pubsub_v1.publisher').setLevel(logging.CRITICAL)
 
-ZIP_LOCATION = os.path.join(tempfile.gettempdir(), 'lithops_gcp.zip')
+ZIP_LOCATION = os.path.join(config.TEMP, 'lithops_gcp.zip')
 SCOPES = ('https://www.googleapis.com/auth/cloud-platform',
           'https://www.googleapis.com/auth/pubsub')
 FUNCTIONS_API_VERSION = 'v1'
@@ -126,7 +125,21 @@ class GCPFunctionsBackend:
     def _get_default_runtime_image_name(self):
         return 'python' + version_str(sys.version_info)
 
-    def _create_handler_zip(self):
+    def _get_runtime_requirements(self, runtime_name):
+        if runtime_name in gcp_config.DEFAULT_RUNTIMES:
+            return gcp_config.DEFAULT_REQUIREMENTS
+        else:
+            user_runtimes = self.internal_storage.storage.list_keys(self.internal_storage.bucket,
+                                                                    prefix=gcp_config.USER_RUNTIMES_PREFIX)
+            user_runtimes_keys = {runtime.split('/', 1)[1]: runtime for runtime in user_runtimes}
+            if runtime_name in user_runtimes_keys:
+                reqs = self.internal_storage.get_data(key=user_runtimes_keys[runtime_name]).decode('utf-8')
+                return reqs.splitlines()
+            else:
+                raise Exception('Runtime {} does not exist. '
+                                'Available runtimes: {}'.format(runtime_name, gcp_config.DEFAULT_RUNTIMES + user_runtimes))
+
+    def _create_handler_zip(self, runtime_name):
         logger.debug("Creating function handler zip in {}".format(ZIP_LOCATION))
 
         def add_folder_to_zip(zip_file, full_dir_path, sub_dir=''):
@@ -137,17 +150,25 @@ class GCPFunctionsBackend:
                 elif os.path.isdir(full_path) and '__pycache__' not in full_path:
                     add_folder_to_zip(zip_file, full_path, os.path.join(sub_dir, file))
 
+        # Get runtime requirements
+        runtime_requirements = self._get_runtime_requirements(runtime_name)
+        requirements_file_path = os.path.join(config.TEMP, '{}_requirements.txt'.format(runtime_name))
+        with open(requirements_file_path, 'w') as reqs_file:
+            for req in runtime_requirements:
+                reqs_file.write('{}\n'.format(req))
+
         try:
             with zipfile.ZipFile(ZIP_LOCATION, 'w') as lithops_zip:
+                # Add Lithops entryfile to zip archive
                 current_location = os.path.dirname(os.path.abspath(__file__))
-                module_location = os.path.dirname(os.path.abspath(lithops.__file__))
                 main_file = os.path.join(current_location, 'entry_point.py')
                 lithops_zip.write(main_file, 'main.py', zipfile.ZIP_DEFLATED)
-                requirements_file_loc = os.path.join(tempfile.gettempdir(), 'requirements.txt')
-                with open(requirements_file_loc, 'w') as req_file:
-                    for req in gcp_config.REQUIREMENTS:
-                        req_file.write('{}\n'.format(req))
-                lithops_zip.write(requirements_file_loc, 'requirements.txt', zipfile.ZIP_DEFLATED)
+
+                # Add runtime requirements.txt to zip archive
+                lithops_zip.write(requirements_file_path, 'requirements.txt', zipfile.ZIP_DEFLATED)
+
+                # Add Lithops to zip archive
+                module_location = os.path.dirname(os.path.abspath(lithops.__file__))
                 add_folder_to_zip(lithops_zip, module_location)
         except Exception as e:
             raise Exception('Unable to create Lithops package: {}'.format(e))
@@ -156,16 +177,17 @@ class GCPFunctionsBackend:
         logger.debug("Creating function {} - Memory: {} Timeout: {} Trigger: {}".format(runtime_name,
                                                                                         memory, timeout, trigger))
         default_location = self._full_default_location()
-        function_location = self._full_function_location(
-            self._format_action_name(runtime_name, memory))
+        function_location = self._full_function_location(self._format_action_name(runtime_name, memory))
         bin_name = self._format_action_name(runtime_name, memory) + '_bin.zip'
         self.internal_storage.put_data(bin_name, code)
+
+        python_runtime_ver = 'python{}'.format(version_str(sys.version_info))
 
         cloud_function = {
             'name': function_location,
             'description': self.package,
             'entryPoint': 'main',
-            'runtime': runtime_name.lower().replace('.', ''),
+            'runtime': python_runtime_ver.lower().replace('.', ''),
             'timeout': str(timeout) + 's',
             'availableMemoryMb': memory,
             'serviceAccountEmail': self.service_account,
@@ -209,11 +231,19 @@ class GCPFunctionsBackend:
         # Delete runtime bin archive from storage
         self.internal_storage.storage.delete_object(self.internal_storage.bucket, bin_name)
 
-    def build_runtime(self):
-        pass
+    def build_runtime(self, runtime_name, requirements_file):
+        runtime_python_ver = 'python{}'.format(version_str(sys.version_info))
+        if runtime_python_ver not in gcp_config.DEFAULT_RUNTIMES:
+            raise Exception('Runtime {} is not available for GCP Functions, '
+                            'please use one of {}'.format(runtime_python_ver, gcp_config.DEFAULT_RUNTIMES))
+
+        with open(requirements_file, 'r') as req_file:
+            requirements = req_file.read()
+
+        self.internal_storage.put_data('/'.join([gcp_config.USER_RUNTIMES_PREFIX, runtime_name]), requirements)
 
     def update_runtime(self, runtime_name, code, memory=3008, timeout=900):
-        pass
+        raise NotImplementedError()
 
     def create_runtime(self, runtime_name, memory, timeout=60):
         logger.debug("Creating runtime {} - Memory: {} Timeout: {}".format(runtime_name, memory, timeout))
@@ -223,24 +253,21 @@ class GCPFunctionsBackend:
 
         # Create topic
         topic_name = self._format_topic_name(runtime_name, memory)
+        topic_list_request = self.publisher_client.list_topics(request={'project': 'projects/{}'.format(self.project)})
         topic_location = self._full_topic_location(topic_name)
-        try:
-            # Try getting topic config
-            self.publisher_client.get_topic(topic=topic_location)
-            # If no exception is raised, then the topic exists
+        topics = [topic.name for topic in topic_list_request]
+        if topic_location in topics:
             logger.info("Topic {} already exists - Restarting queue...".format(topic_location))
             self.publisher_client.delete_topic(topic=topic_location)
-        except google.api_core.exceptions.GoogleAPICallError:
-            pass
         logger.debug("Creating topic {}...".format(topic_location))
         self.publisher_client.create_topic(name=topic_location)
 
         # Create function
-        self._create_handler_zip()
+        self._create_handler_zip(runtime_name)
         with open(ZIP_LOCATION, "rb") as action_zip:
             action_bin = action_zip.read()
 
-        self._create_function(runtime_name, memory, action_bin, timeout=timeout, trigger='Pub/Sub')
+        self._create_function(runtime_name, memory, code=action_bin, timeout=timeout, trigger='Pub/Sub')
 
         return runtime_meta
 
@@ -357,18 +384,34 @@ class GCPFunctionsBackend:
                 return json.dumps(runtime_meta)
         """
         logger.debug('Generating runtime meta for {}...'.format(runtime_name))
-        action_location = os.path.join(tempfile.gettempdir(), 'extract_preinstalls_gcp.py')
-        with open(action_location, 'w') as f:
-            f.write(textwrap.dedent(action_code))
 
-        modules_zip_action = os.path.join(tempfile.gettempdir(), 'extract_preinstalls_gcp.zip')
-        with zipfile.ZipFile(modules_zip_action, 'w') as extract_modules_zip:
-            extract_modules_zip.write(action_location, 'main.py')
-            extract_modules_zip.close()
-        with open(modules_zip_action, 'rb') as modules_zip:
-            action_code = modules_zip.read()
+        # Get runtime requirements
+        runtime_requirements = self._get_runtime_requirements(runtime_name)
+        requirements_file_path = os.path.join(config.TEMP, '{}_requirements.txt'.format(runtime_name))
+        with open(requirements_file_path, 'w') as reqs_file:
+            for req in runtime_requirements:
+                reqs_file.write('{}\n'.format(req))
 
-        self._create_function(runtime_name, 128, action_code, trigger='HTTP')
+        extract_modules_zip_path = os.path.join(config.TEMP, 'extract_modules_gcp.zip')
+        try:
+            with zipfile.ZipFile(extract_modules_zip_path, 'w') as extract_modules_zip:
+
+                # Add action_code entrypoint as main.py to zip archive
+                extract_modules_entrypoint_filename = os.path.join(config.TEMP, 'extract_modules_entrypoint.py')
+                with open(extract_modules_entrypoint_filename, 'w') as extract_mods_entrypoint_f:
+                    extract_mods_entrypoint_f.write(textwrap.dedent(action_code))
+                extract_modules_zip.write(extract_modules_entrypoint_filename, 'main.py', zipfile.ZIP_DEFLATED)
+
+                # Add runtime requirements.txt to zip archive
+                extract_modules_zip.write(requirements_file_path, 'requirements.txt', zipfile.ZIP_DEFLATED)
+
+        except Exception as e:
+            raise Exception('Unable to create Lithops extract_modules package: {}'.format(e))
+
+        with open(extract_modules_zip_path, 'rb') as modules_zip:
+            function_zip_bin = modules_zip.read()
+
+        self._create_function(runtime_name, 128, function_zip_bin, trigger='HTTP')
 
         logger.debug("Extracting Python modules list from: {}".format(runtime_name))
         try:
